cleaning up and bug fix

+++ src/js/api/posts/postFeed.mjs
@@ -165,7 +165,9 @@ ${commentsHtml}
 
     feedContainer.append(card);
 
-    deletePostListener(id);
-    updatePostListener(id);
+    if (currentUser === authorName) {
+      deletePostListener(id);
+      updatePostListener(id);
+    }
   });
 }
